+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Move Coast-RP to new data catalog.
 - Add heat conductivity to deeper soil layers (still missing influence of water).
 - Consider soil heat flux in pennmann-monteith.
+- Turn of sensible and turbulent heat fluxes in case there is snow.
 
 To support this version:
 - Re-run `setup_hydrography`: `geb update -b build.yml::setup_hydrography`